server/requirements.txt
coloredlogs==15.0.1
fastapi==0.110.0
websockets==12.0
openai==1.14.2
pyyaml==6.0.1
vllm==0.3.3
transformers==4.38.2
black==24.3.0
flake8==7.0.0